refactor: move waterAllPlots to its own file
