+++ src/game-logic/reducers/field.js
@@ -41,16 +41,9 @@ import { incrementPlotContentAge } from './incrementPlotContentAge'
 import { resetWasShoveled } from './resetWasShoveled'
 import { applyPrecipitation } from './applyPrecipitation'
 import { showNotification } from './showNotification'
-import { waterField } from './waterField'
 
 const { FERTILIZE, OBSERVE, SET_SCARECROW, SET_SPRINKLER } = fieldMode
 const { GROWN } = cropLifeStage
-
-/**
- * @param {farmhand.state} state
- * @returns {farmhand.state}
- */
-export const waterAllPlots = state => waterField(state)
 
 /**
  * @param {?farmhand.plotContent} plotContent
