+++ src/game-logic/reducers/index.js
@@ -114,6 +114,7 @@ export * from './resetWasShoveled'
 export * from './applyPrecipitation'
 export * from './applyCrows'
 export * from './waterField'
+export * from './waterAllPlots'
 
 ///////////////////////////////////////////////////////////
 //
