+++ src/game-logic/reducers/index.test.js
@@ -2720,24 +2720,6 @@ describe('waterPlot', () => {
   })
 })
 
-describe('waterAllPlots', () => {
-  test('sets wasWateredToday to true for all plots', () => {
-    const { field } = fn.waterAllPlots({
-      field: [
-        [
-          testCrop({ itemId: 'sample-crop-1' }),
-          testCrop({ itemId: 'sample-crop-2' }),
-        ],
-        [testCrop({ itemId: 'sample-crop-3' })],
-      ],
-    })
-
-    expect(field[0][0].wasWateredToday).toBe(true)
-    expect(field[0][1].wasWateredToday).toBe(true)
-    expect(field[1][0].wasWateredToday).toBe(true)
-  })
-})
-
 describe('purchaseCombine', () => {
   test('updates purchasedCombine', () => {
     const { purchasedCombine } = fn.purchaseCombine({}, 1)
