+++ src/game-logic/reducers/waterAllPlots.js
@@ -1,0 +1,8 @@
+import { waterField } from './waterField'
+
+// TODO: Remove this and just use waterField directly.
+/**
+ * @param {farmhand.state} state
+ * @returns {farmhand.state}
+ */
+export const waterAllPlots = state => waterField(state)
